+++ include/mujoco/mujoco.h
@@ -460,6 +460,11 @@ MJAPI void mj_rne(const mjModel* m, mjData* d, int flg_acc, mjtNum* result);
 // RNE with complete data: compute cacc, cfrc_ext, cfrc_int.
 MJAPI void mj_rnePostConstraint(const mjModel* m, mjData* d);
 
+// Return the maximum number of contacts that can be generated between two geoms.
+// If has_margin is -1, then the margin is pulled from the model, otherwise if has_margin > 0
+// indicates that the geoms have a positive margin.
+MJAPI int mj_maxContact(const mjModel* m, int g1, int g2, int has_margin);
+
 // Run collision detection.
 MJAPI void mj_collision(const mjModel* m, mjData* d);